chore: improved the rest apis for organization - Better Dockerfiles for faster builds - Fixed the Schema Conflict for RegisterationResponse - Fixed Refresh Guard - Add Dtos of Memberships - Add Dtos of Organization Responses - Add Centralized PaginationResponseDto - Error Logging while the token verification - Add ApiResponse and File Upload Validators for OrganizationController

+++ src/auth/auth.service.ts
@@ -1,6 +1,7 @@
 import {
   ConflictException,
   Injectable,
+  NotFoundException,
   UnauthorizedException,
 } from '@nestjs/common';
 
@@ -124,10 +125,14 @@ export class AuthService {
     return { accessToken, refreshToken };
   }
 
-  async logout(sessionId: string) {
+  async logout(refreshToken?: string) {
+    if (!refreshToken) return;
+
+    const payload = await this.tokenService.verifyRefreshToken(refreshToken);
+
     await this.prismaService.refreshToken.deleteMany({
       where: {
-        id: sessionId,
+        id: payload.sessionId,
       },
     });
   }
@@ -157,7 +162,15 @@ export class AuthService {
   }
 
   async getMe(id: string) {
-    return this.userService.findUserById(id);
+    const user = await this.userService.findUserById(id);
+    if (!user) throw new NotFoundException();
+
+    return {
+      id: user.id,
+      name: user.name,
+      email: user.email,
+      avatarUrl: user.avatarUrl,
+    };
   }
 
   async updateEmail(id: string, newEmail: string) {
